add protobuf csproj and move internal socket classes to own folder
