+++ Nakama.sln
@@ -11,9 +11,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{4CBC1D9A
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nakama.Tests", "tests\Nakama.Tests\Nakama.Tests.csproj", "{AFD0AF63-EA45-49A6-8889-3E65A48F0521}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "protobuf", "protobuf", "{B2CEABE1-9D77-4C96-8628-62B327F3BBAF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nakama.Protobuf", "protobuf\Nakama.Protobuf\Nakama.Protobuf.csproj", "{B4CE3BC8-FE71-4934-875E-FD38505A8C59}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nakama.Protobuf", "src\Nakama.Protobuf\Nakama.Protobuf.csproj", "{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -52,22 +50,22 @@ Global
 		{AFD0AF63-EA45-49A6-8889-3E65A48F0521}.Release|x64.Build.0 = Release|Any CPU
 		{AFD0AF63-EA45-49A6-8889-3E65A48F0521}.Release|x86.ActiveCfg = Release|Any CPU
 		{AFD0AF63-EA45-49A6-8889-3E65A48F0521}.Release|x86.Build.0 = Release|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Debug|x64.Build.0 = Debug|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Debug|x86.Build.0 = Debug|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Release|x64.ActiveCfg = Release|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Release|x64.Build.0 = Release|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Release|x86.ActiveCfg = Release|Any CPU
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59}.Release|x86.Build.0 = Release|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Debug|x64.Build.0 = Debug|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Debug|x86.Build.0 = Debug|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Release|x64.ActiveCfg = Release|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Release|x64.Build.0 = Release|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Release|x86.ActiveCfg = Release|Any CPU
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{10E81453-27EC-4D8E-82A7-DEDEFABBD648} = {651F309B-3615-44A7-A5B4-832B4ACE2258}
 		{AFD0AF63-EA45-49A6-8889-3E65A48F0521} = {4CBC1D9A-0727-4D31-A3BC-484EF98E7617}
-		{B4CE3BC8-FE71-4934-875E-FD38505A8C59} = {B2CEABE1-9D77-4C96-8628-62B327F3BBAF}
+		{C9CECF38-A4B4-4B91-9ADD-1F37ACA78E2D} = {651F309B-3615-44A7-A5B4-832B4ACE2258}
 	EndGlobalSection
 EndGlobal
