+++ src/Nakama/IChannel.cs
@@ -15,9 +15,8 @@
  */
 
 using System.Collections.Generic;
-using System.Runtime.Serialization;
 
-namespace Nakama
+namespace Nakama.Socket
 {
     /// <summary>
     /// A chat channel on the server.
@@ -58,51 +57,5 @@ namespace Nakama
         /// The ID of the second DM user, or an empty string if this message was not sent through a DM chat.
         /// </summary>
         string UserIdTwo { get; }
-    }
-
-    /// <inheritdoc cref="IChannel"/>
-    internal class Channel : IChannel
-    {
-        [DataMember(Name="id"), Preserve]
-        public string Id { get; set; }
-
-        public IEnumerable<IUserPresence> Presences => _presences ?? UserPresence.NoPresences;
-        [DataMember(Name="presences"), Preserve]
-        public List<UserPresence> _presences { get; set; }
-
-        public IUserPresence Self => _self;
-        [DataMember(Name="self"), Preserve]
-        public UserPresence _self { get; set; }
-
-        [DataMember(Name="room_name"), Preserve]
-        public string RoomName { get; set; }
-
-        [DataMember(Name="group_id"), Preserve]
-        public string GroupId { get; set; }
-
-        [DataMember(Name="user_id_one"), Preserve]
-        public string UserIdOne { get; set; }
-
-        [DataMember(Name="user_id_two"), Preserve]
-        public string UserIdTwo { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (!(obj is Channel item))
-            {
-                return false;
-            }
-            return Equals(item);
-        }
-
-        private bool Equals(IChannel other) => string.Equals(Id, other.Id);
-
-        public override int GetHashCode() => Id != null ? Id.GetHashCode() : 0;
-
-        public override string ToString()
-        {
-            var presences = string.Join(", ", Presences);
-            return $"Channel(Id='{Id}', Presences=[{presences}], Self={Self}, RoomName='{RoomName}', GroupId='{GroupId}', UserIdOne='{UserIdOne}', UserIdTwo='{UserIdTwo}')";
-        }
     }
 }
