+++ src/Nakama/IChannelPresenceEvent.cs
@@ -15,7 +15,6 @@
  */
 
 using System.Collections.Generic;
-using System.Runtime.Serialization;
 
 namespace Nakama
 {
@@ -58,39 +57,5 @@ namespace Nakama
         /// The ID of the second DM user, or an empty string if this message was not sent through a DM chat.
         /// </summary>
         string UserIdTwo { get; }
-    }
-
-    /// <inheritdoc cref="IChannelPresenceEvent"/>
-    internal class ChannelPresenceEvent : IChannelPresenceEvent
-    {
-        [DataMember(Name="channel_id"), Preserve]
-        public string ChannelId { get; set; }
-
-        public IEnumerable<IUserPresence> Joins => _joins ?? new List<UserPresence>(0);
-        [DataMember(Name="joins"), Preserve]
-        public List<UserPresence> _joins { get; set; }
-
-        public IEnumerable<IUserPresence> Leaves => _leaves ?? new List<UserPresence>(0);
-        [DataMember(Name="leaves"), Preserve]
-        public List<UserPresence> _leaves { get; set; }
-
-        [DataMember(Name="room_name"), Preserve]
-        public string RoomName { get; set; }
-
-        [DataMember(Name="group_id"), Preserve]
-        public string GroupId { get; set; }
-
-        [DataMember(Name="user_id_one"), Preserve]
-        public string UserIdOne { get; set; }
-
-        [DataMember(Name="user_id_two"), Preserve]
-        public string UserIdTwo { get; set; }
-
-        public override string ToString()
-        {
-            var joins = string.Join(",", Joins);
-            var leaves = string.Join(",", Leaves);
-            return $"ChannelPresenceEvent(ChannelId='{ChannelId}', Joins=[{joins}], Leaves=[{leaves}], RoomName='{RoomName}', GroupId='{GroupId}', UserIdOne='{UserIdOne}', UserIdTwo='{UserIdTwo}')";
-        }
     }
 }
