+++ src/Nakama/IMatchState.cs
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The Nakama Authors
+ * Copyright 2020 The Nakama Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-using System;
-using System.Runtime.Serialization;
 
 namespace Nakama
 {
@@ -46,27 +43,5 @@ namespace Nakama
         /// Information on the user who sent the state change.
         /// </summary>
         IUserPresence UserPresence { get; }
-    }
-
-    /// <inheritdoc cref="IMatchState"/>
-    internal class MatchState : IMatchState
-    {
-        private static readonly byte[] NoBytes = new byte[0];
-
-        [DataMember(Name = "match_id"), Preserve] public string MatchId { get; set; }
-
-        public long OpCode => Convert.ToInt64(_opCode);
-        [DataMember(Name = "op_code"), Preserve] public string _opCode { get; set; }
-
-        public byte[] State => _state == null ? NoBytes :  Convert.FromBase64String(_state);
-        [DataMember(Name = "data"), Preserve] public string _state { get; set; }
-
-        public IUserPresence UserPresence => _userPresence;
-        [DataMember(Name = "presence"), Preserve] public UserPresence _userPresence { get; set; }
-
-        public override string ToString()
-        {
-            return $"MatchState(MatchId='{MatchId}', OpCode={OpCode}, State='{_state}', UserPresence={UserPresence})";
-        }
     }
 }
