+++ src/Nakama/IMatchmakerMatched.cs
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The Nakama Authors
+ * Copyright 2020 The Nakama Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,6 @@
  */
 
 using System.Collections.Generic;
-using System.Runtime.Serialization;
 
 namespace Nakama
 {
@@ -72,51 +71,5 @@ namespace Nakama
         /// The string properties which this user asked to matchmake with.
         /// </summary>
         IDictionary<string, string> StringProperties { get; }
-    }
-
-    /// <inheritdoc cref="IMatchmakerMatched"/>
-    internal class MatchmakerMatched : IMatchmakerMatched
-    {
-        [DataMember(Name = "match_id"), Preserve] public string MatchId { get; set; }
-
-        [DataMember(Name = "ticket"), Preserve] public string Ticket { get; set; }
-
-        [DataMember(Name = "token"), Preserve] public string Token { get; set; }
-
-        public IEnumerable<IMatchmakerUser> Users => _users ?? new List<MatchmakerUser>(0);
-        [DataMember(Name = "users"), Preserve] public List<MatchmakerUser> _users { get; set; }
-
-        public IMatchmakerUser Self => _self;
-        [DataMember(Name = "self"), Preserve] public MatchmakerUser _self { get; set; }
-
-        public override string ToString()
-        {
-            var users = string.Join(", ", Users);
-            return
-                $"MatchmakerMatched(MatchId='{MatchId}', Ticket='{Ticket}', Token='{Token}', Users=[{users}], Self={Self})";
-        }
-    }
-
-    /// <inheritdoc cref="IMatchmakerUser"/>
-    internal class MatchmakerUser : IMatchmakerUser
-    {
-        public IDictionary<string, double> NumericProperties => _numericProperties ?? new Dictionary<string, double>();
-
-        [DataMember(Name = "numeric_properties"), Preserve]
-        public Dictionary<string, double> _numericProperties { get; set; }
-
-        public IUserPresence Presence => _presence;
-        [DataMember(Name = "presence"), Preserve] public UserPresence _presence { get; set; }
-
-        public IDictionary<string, string> StringProperties => _stringProperties ?? new Dictionary<string, string>();
-
-        [DataMember(Name = "string_properties"), Preserve]
-        public Dictionary<string, string> _stringProperties { get; set; }
-
-        public override string ToString()
-        {
-            return
-                $"MatchmakerUser(NumericProperties={NumericProperties}, Presence={Presence}, StringProperties={StringProperties})";
-        }
     }
 }
