+++ src/Nakama/IStreamPresenceEvent.cs
@@ -90,59 +90,4 @@ namespace Nakama
         /// </summary>
         string Subject { get; }
     }
-
-    /// <inheritdoc cref="IStreamPresenceEvent"/>
-    internal class StreamPresenceEvent : IStreamPresenceEvent
-    {
-        public IEnumerable<IUserPresence> Leaves => _leaves ?? UserPresence.NoPresences;
-        [DataMember(Name = "leaves"), Preserve] public List<UserPresence> _leaves { get; set; }
-
-        public IEnumerable<IUserPresence> Joins => _joins ?? UserPresence.NoPresences;
-        [DataMember(Name = "joins"), Preserve] public List<UserPresence> _joins { get; set; }
-
-        public IStream Stream => _stream;
-        [DataMember(Name = "stream"), Preserve] public Stream _stream { get; set; }
-
-        public override string ToString()
-        {
-            var leaves = string.Join(", ", Leaves);
-            var joins = string.Join(", ", Joins);
-            return $"StreamPresenceEvent(Leaves=[{leaves}], Joins=[{joins}], Stream={Stream})";
-        }
-    }
-
-    /// <inheritdoc cref="IStreamState"/>
-    internal class StreamState : IStreamState
-    {
-        public IUserPresence Sender => _sender;
-        [DataMember(Name = "sender"), Preserve] public UserPresence _sender { get; set; }
-
-        public string State => _state;
-        [DataMember(Name = "data"), Preserve] public string _state { get; set; }
-
-        public IStream Stream => _stream;
-        [DataMember(Name = "stream"), Preserve] public Stream _stream { get; set; }
-
-        public override string ToString()
-        {
-            return $"StreamState(Sender={Sender}, State='{_state}', Stream={Stream})";
-        }
-    }
-
-    /// <inheritdoc cref="IStream"/>
-    internal class Stream : IStream
-    {
-        [DataMember(Name = "descriptor"), Preserve] public string Descriptor { get; set; }
-
-        [DataMember(Name = "label"), Preserve] public string Label { get; set; }
-
-        [DataMember(Name = "mode"), Preserve] public int Mode { get; set; }
-
-        [DataMember(Name = "subject"), Preserve] public string Subject { get; set; }
-
-        public override string ToString()
-        {
-            return $"Stream(Descriptor='{Descriptor}', Label='{Label}', Mode={Mode}, Subject='{Subject}')";
-        }
-    }
 }
