+++ src/Nakama/SocketInternal/ChannelLeaveMessage.cs
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The Nakama Authors
+ * Copyright 2020 The Nakama Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,25 +16,19 @@
 
 using System.Runtime.Serialization;
 
-namespace Nakama
+namespace Nakama.SocketInternal
 {
     /// <summary>
-    /// Update a chat message which has been sent to a channel.
+    /// A leave message to a chat channel.
     /// </summary>
-    internal class ChannelUpdateMessage
+    public class ChannelLeaveMessage
     {
         [DataMember(Name="channel_id"), Preserve]
         public string ChannelId { get; set; }
 
-        [DataMember(Name="message_id"), Preserve]
-        public string MessageId { get; set; }
-
-        [DataMember(Name="content"), Preserve]
-        public string Content { get; set; }
-
         public override string ToString()
         {
-            return $"ChannelUpdateMessage(ChannelId='{ChannelId}', MessageId='{MessageId}', Content='{Content}')";
+            return $"ChannelLeaveMessage(ChannelId='{ChannelId}')";
         }
     }
 }
