+++ src/Nakama/SocketInternal/MatchCreateMessage.cs
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The Nakama Authors
+ * Copyright 2020 The Nakama Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,16 @@
  * limitations under the License.
  */
 
-using System.Runtime.Serialization;
-
-namespace Nakama
+namespace Nakama.SocketInternal
 {
     /// <summary>
-    /// Update the status of the current user.
+    /// A create message for a match on the server.
     /// </summary>
-    internal class StatusUpdateMessage
+    public class MatchCreateMessage
     {
-        [DataMember(Name="status"), Preserve]
-        public string Status { get; set; }
-
         public override string ToString()
         {
-            return $"StatusUpdateMessage(Status='{Status}')";
+            return "MatchCreateMessage()";
         }
     }
 }
