+++ src/Nakama/SocketInternal/MatchJoinMessage.cs
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The Nakama Authors
+ * Copyright 2020 The Nakama Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,22 +17,25 @@
 using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-namespace Nakama
+namespace Nakama.SocketInternal
 {
     /// <summary>
-    /// A logical error received on the WebSocket connection.
+    /// A join message for a match on the server.
     /// </summary>
-    internal class WebSocketErrorMessage
+    public class MatchJoinMessage
     {
-        [DataMember(Name = "code"), Preserve] public int Code { get; set; }
+        [DataMember(Name="match_id"), Preserve]
+        public string MatchId { get; set; }
 
-        [DataMember(Name = "context"), Preserve] public Dictionary<string, string> Context { get; set; }
+        [DataMember(Name="token"), Preserve]
+        public string Token { get; set; }
 
-        [DataMember(Name = "message"), Preserve] public string Message { get; set; }
+        [DataMember(Name="metadata"), Preserve]
+        public IDictionary<string, string> Metadata { get; set; }
 
         public override string ToString()
         {
-            return $"WebSocketErrorMessage(Code={Code}, Context={Context}, Message='{Message}')";
+            return $"MatchJoinMessage(MatchId='{MatchId}', Token='{Token}, Metadata='{Metadata}')";
         }
     }
 }
