+++ src/Nakama/SocketInternal/MatchmakerAddMessage.cs
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The Nakama Authors
+ * Copyright 2020 The Nakama Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,12 +17,12 @@
 using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-namespace Nakama
+namespace Nakama.SocketInternal
 {
     /// <summary>
     /// Add the user to the matchmaker pool with properties.
     /// </summary>
-    internal class MatchmakerAddMessage
+    public class MatchmakerAddMessage
     {
         [DataMember(Name = "max_count"), Preserve] public int MaxCount { get; set; }
