+++ src/Nakama/SocketInternal/StatusFollowMessage.cs
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The Nakama Authors
+ * Copyright 2020 The Nakama Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,24 +14,25 @@
  * limitations under the License.
  */
 
+using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-namespace Nakama
+namespace Nakama.SocketInternal
 {
     /// <summary>
-    /// Remove a message from a chat channel.
+    /// Follow one or more other users for status updates.
     /// </summary>
-    internal class ChannelRemoveMessage
+    public class StatusFollowMessage
     {
-        [DataMember(Name="channel_id"), Preserve]
-        public string ChannelId { get; set; }
+        [DataMember(Name = "user_ids"), Preserve] public List<string> UserIds { get; set; }
 
-        [DataMember(Name="message_id"), Preserve]
-        public string MessageId { get; set; }
+        [DataMember(Name = "usernames"), Preserve] public List<string> Usernames { get; set; }
 
         public override string ToString()
         {
-            return $"ChannelRemoveMessage(ChannelId='{ChannelId}', MessageId='{MessageId}')";
+            var userIds = string.Join(", ", UserIds);
+            var usernames = string.Join(", ", Usernames);
+            return $"StatusFollowMessage(UserIds=[{userIds}],Usernames=[{usernames}])";
         }
     }
 }
