+++ src/Nakama/SocketInternal/StatusUpdateMessage.cs
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The Nakama Authors
+ * Copyright 2020 The Nakama Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,21 @@
  * limitations under the License.
  */
 
-using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-namespace Nakama
+namespace Nakama.SocketInternal
 {
     /// <summary>
-    /// Receive status updates for users.
+    /// Update the status of the current user.
     /// </summary>
-    public interface IStatus
+    public class StatusUpdateMessage
     {
-        /// <summary>
-        /// The status events for the users followed.
-        /// </summary>
-        IEnumerable<IUserPresence> Presences { get; }
+        [DataMember(Name="status"), Preserve]
+        public string Status { get; set; }
+
+        public override string ToString()
+        {
+            return $"StatusUpdateMessage(Status='{Status}')";
+        }
     }
 }
